test(analyzer): fold the contract tests into the loader test file
Review feedback: tests in this repo are arranged by the file they test, so
a standalone test_predefined_recognizer_contract.py is the wrong shape.

Moved into tests/test_recognizers_loader_utils.py, which covers
recognizers_loader_utils.py -- the module whose contract these assert.
RecognizerListLoader is what passes `name`/`supported_language` to every
predefined recognizer, what resolves an entry's class, and what consumes
default_recognizers.yaml. That file already hosts the sibling shipped-YAML
check (test_default_recognizers_yaml_country_code_matches_class), so the
two now sit together instead of in separate files.

The tests themselves are unchanged; the module docstring became a section
header. test_default_recognizers_yaml_country_code_matches_class now reads
the shipped YAML through the module-level constant the move introduced,
rather than re-opening the same file a second time in one module.

Test count is unchanged at 295 (294 pass, 1 skip -- BasicLangExtract, no
optional dependency installed).
